y accel is actually double

+++ PlayerObject.h
@@ -14,7 +14,7 @@ namespace PlayerObject {
 	void __fastcall releaseButtonHook(void* self, void*, void*);
 
 	float* getX(uintptr_t);
-	float* getYAccel(uintptr_t);
+	double* getYAccel(uintptr_t);
 	float* getRotation(uintptr_t);
 	float* getSpriteRotation(uintptr_t);
 }
